feat #1037: Mejoras UI
- Refactor entradas con elementos.
