+++ src/app/pages/entradas/aprovechamientos/aprovechamientos.component.ts
@@ -8,8 +8,8 @@ import { TransaccionEntrada } from '../../../@core/data/models/entrada/entrada';
 import { TranslateService } from '@ngx-translate/core';
 import { EntradaHelper } from '../../../helpers/entradas/entradaHelper';
 import { MatPaginator, MatTableDataSource } from '@angular/material';
-import { MovimientosHelper } from '../../../helpers/movimientos/movimientosHelper';
 import { CommonEntradas } from '../CommonEntradas';
+import { CommonElementos } from '../CommonElementos';
 
 @Component({
   selector: 'ngx-aprovechamientos',
@@ -41,18 +41,18 @@ export class AprovechamientosComponent implements OnInit {
 
   constructor(
     private common: CommonEntradas,
-    private movimientos: MovimientosHelper,
+    private commonElementos: CommonElementos,
     private entradasHelper: EntradaHelper,
     private pUpManager: PopUpManager,
     private fb: FormBuilder,
     private translate: TranslateService,
   ) {
-    this.displayedColumns = this.common.columnsElementos;
     this.dependenciaSupervisor = '';
   }
 
   ngOnInit() {
-    this.elementosForm = this.common.formElementos;
+    this.displayedColumns = this.commonElementos.columnsElementos;
+    this.elementosForm = this.commonElementos.formElementos;
     this.observacionForm = this.common.formObservaciones;
     this.supervisorForm = this.fb.group({
       supervisorCtrl: ['', Validators.required],
@@ -63,33 +63,17 @@ export class AprovechamientosComponent implements OnInit {
   }
 
   addElemento() {
-    (this.elementosForm.get('elementos') as FormArray).push(this.elemento);
+    const form = this.commonElementos.elemento;
+    (this.elementosForm.get('elementos') as FormArray).push(form);
     this.dataSource.data = this.dataSource.data.concat({});
-  }
-
-  get elemento(): FormGroup {
-    const form = this.common.elemento;
     this.cambiosPlaca(form.get('Placa').valueChanges);
-    return form;
-  }
-
-  getActualIndex(index: number) {
-    return index + this.paginator.pageSize * this.paginator.pageIndex;
-  }
-
-  removeElemento(index: number) {
-    index = this.paginator.pageIndex > 0 ? index + (this.paginator.pageIndex * this.paginator.pageSize) : index;
-    (this.elementosForm.get('elementos') as FormArray).removeAt(index);
-    const data = this.dataSource.data;
-    data.splice(index, 1);
-    this.dataSource.data = data;
   }
 
   private cambiosPlaca(valueChanges: Observable<any>) {
     valueChanges.pipe(
       debounceTime(250),
       distinctUntilChanged(),
-      switchMap((val) => this.common.loadElementos(val)),
+      switchMap((val) => this.commonElementos.loadElementos(val)),
     ).subscribe((response: any) => {
       this.elementos = response.queryOptions;
     });
@@ -112,7 +96,6 @@ export class AprovechamientosComponent implements OnInit {
             map(val => typeof val === 'string' ? val : this.muestraSupervisor(val)),
             map(nombre => this.filtroSupervisores(nombre)),
           );
-        // console.log({supervisores: this.Supervisores});
         this.cargando_supervisores = false;
       }
     });
